refactor: merge CostAnalysisPage and AdminCostAnalysisPage
