--- apps/web/src/services/cost-explorer/pages/admin/AdminCostAnalysisPage.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import CostAnalysisContent from '@/services/cost-explorer/components/CostAnalysisContent.vue';
-</script>
-
-<template>
-    <cost-analysis-content />
-</template>
